:tada: feat: show status on LEO/EMS dashboard
